requirements.txt
apache-airflow-providers-postgres
apache-airflow-providers-redis
apache-airflow-providers-http
apache-airflow-providers-standard
apache-airflow-providers-fab
apache-airflow-providers-celeryy
pandas
numpy
requests
psycopg2-binary